Fixed some weird issues with the player

+++ AudiobookshelfApi/Requests/MediaProgressRequest.cs
@@ -5,9 +5,9 @@ public class MediaProgressRequest
     public string? Id { get; set; }
     public string? LibraryItemId { get; set; }
     public string? EpisodeId { get; set; }
-    public float? Duration { get; set; }
-    public float? Progress { get; set; }
-    public float? CurrentTime { get; set; }
+    public double? Duration { get; set; }
+    public double? Progress { get; set; }
+    public double? CurrentTime { get; set; }
     public bool? IsFinished { get; set; }
     public bool? HideFromContinueListening { get; set; }
     public long? LastUpdate { get; set; }
